captcha_model/requirements.txt
torch
torchvision
efficientnet_pytorch
onnx
onnxruntime
Pillow
numpy
tqdm
tensorboard
